network, p2p, vendor: move vendored p2p/testing under swarm
